requirements.txt

## File: requirements:

```txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# Deep learning
tensorflow==2.13.0
tensorflow-addons==0.21.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2

# Machine learning
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2

# Geospatial & climate
xarray==2023.7.0
netCDF4==1.6.4
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0
cartopy==0.21.1
rioxarray==0.15.0

# Satellite data
earthengine-api==0.1.361
sentinelsat==1.2.1
pystac-client==0.7.2
planetary-computer==1.0.0

# Climate data access
cdsapi==0.6.1
ecmwf-api-client==1.6.3

# Web scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
aiohttp==3.8.5

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
folium==0.14.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.65.0
joblib==1.3.1
dask[complete]==2023.7.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.1.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
